Generalized copy paste functionality
