+++ client/src/components/Table/index.jsx
@@ -123,73 +123,138 @@ export const createCellRenderer = (type, getData, options = {}) => {
   }
 };
 
-const DataTable = (props) => {
-  const {
-    columns,
-    size,
-    total,
-    enableReordering,
-    enableResizing,
-    renderContextMenuCallback,
-    reorderColumnsCallback,
-    resizeColumnCallback,
-    numFrozenColumns,
-    enableGhostCells,
-    copyCallback,
-    selectionCallback,
-    enableRowHeader,
-    selectionMode,
-  } = props;
-  let { rowHeight } = props;
-  const rowsCount = size <= total ? size : total;
-  const handleColumnsReordered = (oldIndex, newIndex, length) => {
-    if (oldIndex === newIndex) {
-      return;
+class DataTable extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.copyCallback = this.copyCallback.bind(this);
+    this.selectionCallback = this.selectionCallback.bind(this);
+
+    this.regionSelectedArray = [];
+  }
+
+  copyCallback() {
+    if (this.regionSelectedArray) {
+      const copyTextToClipboard = (text) => {
+        navigator.clipboard.writeText(text).then(() => {
+        }, (err) => {
+          console.log('Could not copy text: ', err);
+        });
+      };
+      const copyArrayToClipboard = (array) => {
+        let csv = '';
+        array.forEach((row, rowIdx) => {
+          row.forEach((cell, cellIdx) => {
+            csv += (`${cell}`).replace(/[\n\t]+/g, ' ');
+            if (cellIdx + 1 < row.length) csv += '\t';
+          });
+          if (rowIdx + 1 < array.length) csv += '\n';
+        });
+        copyTextToClipboard(csv);
+      };
+
+      copyArrayToClipboard(this.regionSelectedArray);
+    }
+  }
+
+  selectionCallback(sel) {
+    const {
+      getData,
+    } = this.props;
+
+    const selection = sel[0];
+    const rowStart = selection.rows[0];
+    const rowEnd = selection.rows[selection.rows.length - 1];
+    const colStart = selection.cols[0];
+    const colEnd = selection.cols[selection.cols.length - 1];
+
+    const {
+      columns,
+    } = this.props;
+
+    const array = [];
+    for (let row = rowStart; row <= rowEnd; row += 1) {
+      const newRow = [];
+      for (let col = colStart; col <= colEnd; col += 1) {
+        const renderer = columns[col].excelRenderer;
+        if (renderer) {
+          const output = (`${renderer(getData()[row])}`).replace(/[\n\t]+/g, ' ');
+          newRow.push(output);
+        }
+      }
+      array.push(newRow);
     }
 
-    reorderColumnsCallback(Utils.reorderArray(columns, oldIndex, newIndex, length));
-  };
-  rowHeight = rowsCount === 0 ? [] : rowHeight;
-  if (rowsCount < rowHeight.length) {
-    rowHeight = rowHeight.slice(0, rowsCount);
+    this.regionSelectedArray = array;
   }
-  if (rowHeight.length < rowsCount) {
-    const bufferArray = Array(rowsCount - rowHeight.length).fill(32);
-    rowHeight = [...rowHeight, ...bufferArray];
+
+  render() {
+    const {
+      columns,
+      size,
+      total,
+      enableReordering,
+      enableResizing,
+      renderContextMenuCallback,
+      reorderColumnsCallback,
+      resizeColumnCallback,
+      numFrozenColumns,
+      enableGhostCells,
+      enableRowHeader,
+    } = this.props;
+    let { rowHeight } = this.props;
+    const rowsCount = size <= total ? size : total;
+    const handleColumnsReordered = (oldIndex, newIndex, length) => {
+      if (oldIndex === newIndex) {
+        return;
+      }
+
+      reorderColumnsCallback(Utils.reorderArray(columns, oldIndex, newIndex, length));
+    };
+    rowHeight = rowsCount === 0 ? [] : rowHeight;
+    if (rowsCount < rowHeight.length) {
+      rowHeight = rowHeight.slice(0, rowsCount);
+    }
+    if (rowHeight.length < rowsCount) {
+      const bufferArray = Array(rowsCount - rowHeight.length).fill(32);
+      rowHeight = [...rowHeight, ...bufferArray];
+    }
+    const renderColumnHeader = (name, index) => (
+      <div className="tooltipHeader">{intl.get(columns[index].label)} <IconKit size={16} icon={ic_info_outline} /></div>
+    );
+    return (
+      <Table
+        key={shortid.generate()}
+        numRows={rowsCount}
+        numFrozenColumns={numFrozenColumns}
+        enableGhostCells={enableGhostCells}
+        renderMode={RenderMode.BATCH_ON_UPDATE}
+        enableColumnReordering={enableReordering}
+        enableColumnResizing={enableResizing}
+        bodyContextMenuRenderer={renderContextMenuCallback}
+        onColumnsReordered={handleColumnsReordered}
+        onColumnWidthChanged={resizeColumnCallback}
+        rowHeights={rowHeight}
+        getCellClipboardData={(row, col) => `(${row}, ${col})`}
+        columnWidths={columns.map(c => c.columnWidth)}
+        enableRowHeader={enableRowHeader}
+        selectionMode={SelectionModes.ROWS_AND_CELLS}
+        // selectionMode={SelectionModes.NONE}
+        onCopy={this.copyCallback}
+        onSelection={this.selectionCallback}
+      >
+        { columns.map(definition => (
+          <Column
+            id={definition.key}
+            name={definition.description ? definition.description : intl.get(definition.label)}
+            cellRenderer={definition.renderer}
+            nameRenderer={definition.description ? renderColumnHeader : null}
+          />
+        )) }
+      </Table>
+    );
   }
-  const renderColumnHeader = (name, index) => (<div className="tooltipHeader">{intl.get(columns[index].label)} <IconKit size={16} icon={ic_info_outline} /></div>);
-  return (
-    <Table
-      key={shortid.generate()}
-      numRows={rowsCount}
-      numFrozenColumns={numFrozenColumns}
-      enableGhostCells={enableGhostCells}
-      renderMode={RenderMode.BATCH_ON_UPDATE}
-      enableColumnReordering={enableReordering}
-      enableColumnResizing={enableResizing}
-      bodyContextMenuRenderer={renderContextMenuCallback}
-      onColumnsReordered={handleColumnsReordered}
-      onColumnWidthChanged={resizeColumnCallback}
-      rowHeights={rowHeight}
-      getCellClipboardData={(row, col) => `(${row}, ${col})`}
-      // onCopy={copyCallback}
-      columnWidths={columns.map(c => c.columnWidth)}
-      enableRowHeader={enableRowHeader}
-      selectionMode={selectionMode}
-      onCopy={copyCallback}
-      onSelection={selectionCallback}
-    >
-      { columns.map(definition => (
-        <Column
-          id={definition.key}
-          name={definition.description ? definition.description : intl.get(definition.label)}
-          cellRenderer={definition.renderer}
-          nameRenderer={definition.description ? renderColumnHeader : null}
-        />
-      )) }
-    </Table>
-  );
-};
+}
 
 DataTable.propTypes = {
   size: PropTypes.number,
@@ -202,11 +267,9 @@ DataTable.propTypes = {
   renderContextMenuCallback: PropTypes.func,
   reorderColumnsCallback: PropTypes.func,
   resizeColumnCallback: PropTypes.func,
-  copyCallback: PropTypes.func,
-  selectionCallback: PropTypes.func,
+  getData: PropTypes.func,
   rowHeight: PropTypes.shape([]),
   enableRowHeader: PropTypes.bool,
-  selectionMode: PropTypes.number,
 };
 
 DataTable.defaultProps = {
@@ -220,11 +283,9 @@ DataTable.defaultProps = {
   renderContextMenuCallback: () => {},
   reorderColumnsCallback: () => {},
   resizeColumnCallback: () => {},
-  copyCallback: null,
-  selectionCallback: null,
+  getData: () => {},
   rowHeight: [],
   enableRowHeader: true,
-  selectionMode: SelectionModes.NONE,
 };
 
 export default DataTable;
